# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/staff/{id}":{"get":{"tags":["staff-controller"],"operationId":"getStaffById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Staff"}}}}}},"put":{"tags":["staff-controller"],"operationId":"updateStaff","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Staff"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Staff"}}}}}},"delete":{"tags":["staff-controller"],"operationId":"deleteStaff","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/rooms/{id}":{"get":{"tags":["room-controller"],"operationId":"getRoomById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Room"}}}}}},"put":{"tags":["room-controller"],"operationId":"updateRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Room"}}}}}},"delete":{"tags":["room-controller"],"operationId":"deleteRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/prescriptions/{id}":{"get":{"tags":["prescription-controller"],"operationId":"getPrescriptionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Prescription"}}}}}},"put":{"tags":["prescription-controller"],"operationId":"updatePrescription","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Prescription"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Prescription"}}}}}},"delete":{"tags":["prescription-controller"],"operationId":"deletePrescription","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/patients/{id}":{"get":{"tags":["patient-controller"],"operationId":"getPatientById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Patient"}}}}}},"put":{"tags":["patient-controller"],"operationId":"updatePatient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Patient"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Patient"}}}}}},"delete":{"tags":["patient-controller"],"operationId":"deletePatient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/hospitalizations/{id}":{"get":{"tags":["hospitalization-controller"],"operationId":"getHospitalizationById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Hospitalization"}}}}}},"put":{"tags":["hospitalization-controller"],"operationId":"updateHospitalization","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Hospitalization"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Hospitalization"}}}}}},"delete":{"tags":["hospitalization-controller"],"operationId":"deleteHospitalization","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/exams/{id}":{"get":{"tags":["exam-controller"],"operationId":"getExamById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Exam"}}}}}},"put":{"tags":["exam-controller"],"operationId":"updateExam","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Exam"}}}}}},"delete":{"tags":["exam-controller"],"operationId":"deleteExam","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/doctors/{id}":{"get":{"tags":["doctor-controller"],"operationId":"getDoctorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Doctor"}}}}}},"put":{"tags":["doctor-controller"],"operationId":"updateDoctor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Doctor"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Doctor"}}}}}},"delete":{"tags":["doctor-controller"],"operationId":"deleteDoctor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/appointments/{id}":{"get":{"tags":["appointment-controller"],"operationId":"getAppointmentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Appointment"}}}}}},"put":{"tags":["appointment-controller"],"operationId":"updateAppointment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Appointment"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Appointment"}}}}}},"delete":{"tags":["appointment-controller"],"operationId":"deleteAppointment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/staff":{"get":{"tags":["staff-controller"],"operationId":"getAllStaff","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Staff"}}}}}}},"post":{"tags":["staff-controller"],"operationId":"createStaff","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Staff"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Staff"}}}}}}},"/api/rooms":{"get":{"tags":["room-controller"],"operationId":"getAllRooms","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Room"}}}}}}},"post":{"tags":["room-controller"],"operationId":"createRoom","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Room"}}}}}}},"/api/prescriptions":{"get":{"tags":["prescription-controller"],"operationId":"getAllPrescriptions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Prescription"}}}}}}},"post":{"tags":["prescription-controller"],"operationId":"createPrescription","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Prescription"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Prescription"}}}}}}},"/api/patients":{"get":{"tags":["patient-controller"],"operationId":"getAllPatients","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Patient"}}}}}}},"post":{"tags":["patient-controller"],"operationId":"createPatient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Patient"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Patient"}}}}}}},"/api/hospitalizations":{"get":{"tags":["hospitalization-controller"],"operationId":"getAllHospitalizations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Hospitalization"}}}}}}},"post":{"tags":["hospitalization-controller"],"operationId":"createHospitalization","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Hospitalization"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Hospitalization"}}}}}}},"/api/exams":{"get":{"tags":["exam-controller"],"operationId":"getAllExams","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Exam"}}}}}}},"post":{"tags":["exam-controller"],"operationId":"createExam","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Exam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Exam"}}}}}}},"/api/doctors":{"get":{"tags":["doctor-controller"],"operationId":"getAllDoctors","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Doctor"}}}}}}},"post":{"tags":["doctor-controller"],"operationId":"createDoctor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Doctor"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Doctor"}}}}}}},"/api/appointments":{"get":{"tags":["appointment-controller"],"operationId":"getAllAppointments","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"}}}}}}},"post":{"tags":["appointment-controller"],"operationId":"createAppointment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Appointment"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Appointment"}}}}}}}},"components":{"schemas":{"Staff":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"role":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"}}},"Room":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"floor":{"type":"integer","format":"int32"},"capacity":{"type":"integer","format":"int32"},"equipment":{"type":"string"}}},"Appointment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateTime":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["SCHEDULED","IN_PROGRESS","COMPLETED","CANCELED"]},"notes":{"type":"string"},"patient":{"$ref":"#/components/schemas/Patient"},"doctor":{"$ref":"#/components/schemas/Doctor"},"room":{"$ref":"#/components/schemas/Room"}}},"Doctor":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"licenseNumber":{"type":"string"},"specialty":{"type":"string","enum":["GENERAL_MEDICINE","CARDIOLOGY","PEDIATRICS","DERMATOLOGY","ORTHOPEDICS","PSYCHIATRY"]},"phone":{"type":"string"},"email":{"type":"string"}}},"Patient":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"gender":{"type":"string","enum":["MALE","FEMALE","OTHER"]},"phone":{"type":"string"},"email":{"type":"string"},"ssn":{"type":"string"},"address":{"type":"string"},"healthCardNumber":{"type":"string"}}},"Prescription":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"medications":{"type":"string"},"instructions":{"type":"string"},"appointment":{"$ref":"#/components/schemas/Appointment"}}},"Hospitalization":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"admissionDate":{"type":"string","format":"date-time"},"dischargeDate":{"type":"string","format":"date-time"},"notes":{"type":"string"},"patient":{"$ref":"#/components/schemas/Patient"},"room":{"$ref":"#/components/schemas/Room"}}},"Exam":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"result":{"type":"string"},"examDate":{"type":"string","format":"date-time"},"appointment":{"$ref":"#/components/schemas/Appointment"},"doctor":{"$ref":"#/components/schemas/Doctor"}}}}}}