# hospital-management-backend/src/main/java/com/example/hospitalmanagement/model/Patient.java
package com.example.hospitalmanagement.model;

import lombok.*;
import javax.persistence.*;
import java.time.LocalDate;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class Patient {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false)
    private String name;

    @Column(nullable = false)
    private LocalDate dateOfBirth;

    @Enumerated(EnumType.STRING)
    @Column(nullable = false)
    private Gender gender;

    private String phone;
    private String email;
    private String ssn;
    private String address;

    @Column(nullable = false)
    private String healthCardNumber;
}