# hospital-management-backend/src/main/java/com/example/hospitalmanagement/model/Room.java
package com.example.hospitalmanagement.model;

import lombok.*;
import javax.persistence.*;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class Room {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false)
    private String name;

    private Integer floor;
    private Integer capacity;
    private String equipment;
}